fix(movies-ui): keep save toasts visible
Fixes #1486
